INTERLUDE: turns out I had to refactor the TaskRun data model a bit. We care about NodeRun's, and a TaskRun is a type of NodeRun. An ExternalEventRun is also a type of NodeRun.
Therefore, we renamed TaskRun to NodeRun, and added a oneof field which has info that's specific to Task/ExternalEvent/etc. This mirrors the way that they are handled in the Node (WfSpec/metadata declaration) class. Much cleaner; however, it took a massive refactor. Next up: the TODO's from the last commit.