Adds rudimentary exception handling.
The infrastructure for catching failures is highly similar (and involves a little too much copy-pasta) relative to the Interrupt handling infrastructure. I decided to treat them separately at the cost of more redundant code so that they could be logically distinct from the user point of view.

I hypothesize that the similarities in code required to do catch-exception and handle-interrupt is due to the fact that, in REAL programming, both interrupt handlers and exception handlers involve suspending the current thread in User space and executing some voodoo in Kernel space...translated into workflow semantics, that means that the parent thread gets halted, then the Scheduler schedules a child thread which runs and must complete before the parent can resume.

A note on Failure Handler semantics:
1. If the Failure Handler reaches COMPLETED, then the parent (failed) thread resumes processing by advancing from the next node.
2. If the Failure Handler fails (ERROR state), then the parent thread fails too and the error propagates up (either to the parent of the parent, or if the parent is already the root, then the whole WfRun stops).
3. The Failure Handler is just another child thread, so if you want to mutate some parent variables before continuing with the workflow, you may.
4. You can throw exceptions with specific names. But certain exceptions are reserved, eg VAR_SUB_ERROR, VAR_MUTATION_ERROR, TIMEOUT, TASK_FAILURE, CHILD_FAILURE, VAR_ERROR, TASK_ERROR.
5. You can catch exceptions of specific types.
6. The content of the Failure is passed as INPUT to the handler spec name, just like interrupts.