Refactoring:
- Enhances the Tag class to contain a list of 'Attributes' as well as a 'TagStorageTypePb'.
- Tag class is just a normal storeable and can calculate the objectId based on the attributes and the stored object type.
- PaginatedTagQuery now relies on the list of attributes rather than the raw string.

The above refactors are just to clean up and separate internal layers of abstraction so my own life is easier (and whoever joins as first employee lol).

Next up is to connect the TagStorageType info to the storing of the TagLocalCounter, which should only be stored for the LOCAL_COUNTED storage type.