Fixes issue where we use IQ-based store from within the processor.
Next up: fix issue caused when we:
1. Are about to return a task to a client, so we record the TaskClaimEvent
2. Before that processor processes the TaskClaimEvent, a Rebalance occurs and the task migrates.
3. The AsyncWaiter for that taskClaimEvent just sits forever on the original server, and never gets processed.
4. The calling task worker never gets the reply, so the workers stall until a restart, and tasks never get claimed...yikes.

The simplest solution is to just put a timeout on processing for the TaskClaimEvent (or any Command in the AsyncWaiters), and to schedule a background thread that periodically purges it and calls onError() or onNext(<<some error code>>) to whatever waiter there is.

A smarter solution would be to have an AsyncWaiters apparatus per-partition, and to purge it when the partition/Streams Task is migrated to another instance. We can do that later, but for now let's just get it working.