Merged in LH-23-fix-interrupt-test (pull request #5)
LH-23: Fix Interrupt Test.

* LH-23: Fix Interrupt Test.

Interrupts rely upon comparing ExternalEventId's in the HaltReason and the PendingInterrupt (in the ThreadRun::processExternalEvent() method). Before we created the ObjectId abstract class, all object id's were just Strings, so we used '.equals()'.

However, now it's a regular POJO that we wrote, so we need to add the object id to it.