Merged in fix/improper-store-name-on-central-search (pull request #12)
Fixes an issue that was not apparent when testing on only one server instance.

* Fixes an issue that was not apparent when testing on only one server instance.

When we added a repartitioned store, the InternalScan now requires a store name so that the handler in BackendInternalComms knows which store to request from. We forgot to set that store name, so the handler was trying to look up the "" store, which doesn't exist.

* PR requests


Approved-by: Saúl Piña