Search WfRun By Variables (#177)
This commit allows the user to pass in a `repeated VariableMatch variable_filters` field to the `SearchWfRunRequest`. The behavior is as follows:

- Exact match is performed on the variables.
- The matches use "AND" semantics.

This feature is in the alpha status and requires more work, such as:
- Optimizations to use the appropriate indexes to speed up the query.
- Unit, Integration, and End-to-End tests.

Note that this API is not extensible to more complex queries (with AND, OR, BETWEEN, etc). For those queries, we are designing an additional endpoint which will take in a structured query and return various results as specified by the query.