Fixes Metadata Search Pagination (#567)
Metadata Search Pagination was broken by ambiguity between cluster-scoped and tenant-scoped stores when we re-added the default/ prefix back for the default Tenant.

This PR is a TEMPORARY fix for this issue which hard-codes correct logic to handle metadata searches. The combination of #556 PR and the planned follow-up will remove this hackiness.

This PR also adds an e2e test which verifies that pagination works for metadata search.