ENHANCEMENT: Allows Task Worker to Wait for TaskDef
Metadata is strongly consistent on the write path, but it is eventually consistent on the read path as it must propagate from the leader for the metadata topology to the global store on each of the Core SubTopology instances. This means that if you do a putTaskDef() and then LHTaskWorker#start() in quick succession, the Task Worker can crash with a TaskDef NOT_FOUND error.

This is not dangerous, since we still have consistent write processing and eventual consistency for reads.

To make this less of a pain in the butt, this commit makes the Java Task Worker wait (at most two seconds) until the TaskDef has propagated to the instances. It also adds an e2e test to verify this behavior. Before the modifications to LHTaskWorker.java, I verified that the e2e test failed.