FIX: NPE when accessing variable outside of scope.
When a ThreadSpec attempts to mutate a real variable (but that variable is out of scope for that threadspec) the server threw an NPE and ignored the Command that triggered the advance that resulted in reading that variable.

The proper behavior is for the illegal access to throw an LHVarSubError.

This commit adds a failing e2e test for that case and fixes it. However, the e2e test has apparently discovered another bug related to moving a failed thread from HALTED to ERROR, which will be addressed in subsequent commits.