Creates new Request Context for each onNext in the PollObserver (#717)
Fixes issue after rebalances wherein a `RequestExecutionContext` was created once for the lifetime of a `rpc PollTask`, and a `RequestExecutionContext` opened a Kafka Streams State Store at the beginning of the task worker polling, and a task worker can poll for a long time (hours/days/weeks).

If a kafka streams rebalance occurs during this process, then the server can get some "Invalid State Store Exception"s. This PR fixes this by always keeping the `RequestExecutionContext` fresh on each message.