Minor: Don't wait for ReportTaskRun
There is no need to wait for the processing of the ReportTaskRun command since it does not return anything meaningful to the client. Therefore, we can just return the same google.protobuf.Empty to the client as soon as the Command is safely recorded to Kafka, rather than using the async waiters.

This should help in two ways:
- Reduce the number of things in the AsyncWaiters
- Make rolling restarts appear smoother from the client (task worker) perspective