MINOR: Cache instance name (#792)
In the CommandProcessor, we log the 'config.getLHInstanceName()'. When the LHS_INSTANCE_ID is not set, this causes the 'ConfigBase.java' to do a WARN log of 'LHS_INSTANCE_ID not set'.

This is a short-term fix to just not do that log every time. The long term fix is that the getOrSetDefault(foo, null) should make ConfigBase 'remember' that the key 'foo' has already been asked for, so it only does the WARN log on the first time.