build(server): use tag version on server artifact (#978)
Fixes some issues with the advertised version for lh-server and lhctl docker image.

For lhctl now we pass two variables VERSION and COMMIT which are set as build args and in consequence the resulting lhctl binary will output the correct version when calling lhctl version from the docker image.

For lh-server the version class was adding the PreReleaseIdentifier as empty string, which caused lhctl to print an additional character -. Also after the version replacement occurs during a release, the test was expecting to have a static value 0.0.0-development but during the production build that value would change. So I refactored the test to be flexible enough to validate the version to be published.