fix(server): Adds processing exception handler for core processor (#992)
This PR includes a refactor to the exception handling mechanism for both core and metadata command processors. We want to propagate most of the kafka exceptions (except for RecordTooLargeException) to the Kafka streams internals, so Kafka streams will decide how to handle those exceptions (sometimes killing the stream thread), but we want to avoid poison pill scenarios by adding a log and continue strategy. We plan to integrate this code easily with Kafka Streams 3.9, which will include a better way to handle processing exceptions.

Also, we add a onCancelHandler to the Grpc observers in order to avoid exceptions when asyncwaiters try to close an observer that was previously closed