feat(server): use kip-899 to rebootstrap clients
In testing, I observed that when moving from a 3-node Kafka cluster with mixed brokers and controllers to a Kafka cluster with separate brokers and controllers, all of the Kafka clients got stuck and continued trying to connect to the controllers which no longer had the broker role.

It turns out that KIP-899 indirectly solves this problem by allowing clients to re-bootstrap and discover which servers are the brokers now. When tested with my fix for KAFKA-17455, we are successfully able to continue processing workflows without a crash / TaskCorruptedException when migrating the Kafka Cluster from mixed mode to separate mode.