perf(server): leave consumer group on shutdown by default
There is an undocumented and unfortunate behavior in Kafka Streams such that if:
- One instance joins a group
- Another instance calls KafkaStreams#close() before the Rebalance is over
- then the close() call will block until max.poll.interval.ms

That is really not good. However, the KS community says they have no plans to fix it.

One interesting thing though is that if we send the 'LeaveGroupRequest' then close() returns immediately (as expected). This PR sets the default of the LHS_X_LEAVE_GROUP_ON_SHUTDOWN config to true, so that by default the LH Server Streams Topologies will send a LeaveGroupRequest, and allow for clean shutdowns.

We still leave the configuration there for adventurous users who want to play with it.