perf(server): avoid sending extra writes to rocksdb (#1490)
In the GetableManager#commit() method, we previously put every single GetableToStore into rocksdb. THat's problematic because when we *read* a getable without modifying it, it ends up inside the buffer. This PR reduces the writes to RocksDB by saving the previous protobuf and comparing to the protobuf at the time of commit. This costs a bit more CPU at the server-side but reduces the number of writes to rocksdb.