perf(server): put getables for a single wfrun under same prefix
This backwards-compatible change puts all getable's such as TaskRun's and NodeRun's for a single WfRun onto the same prefix. Doing this will put most of the keys for a WfRun onto the same block in RocksDB (or in adjacent blocks), which optimizes I/O both in compaction and in cold reads.