ci: automatically build and publish snapshot versions into maven repositories.  (#1863)
This PR adds the ability to build and publish snapshots versions every week on Fridays. Also, this PR modifies the current version management in the repository, putting a static version in the gradle.properties that can be associated with every commit.

This PR opens the possibility to optimize the current internal release process. This process is going to be revised accordingly.