fix(server): dodge tombstones on rehydration (#1902)
See this documentation in RocksDB: https://github.com/facebook/rocksdb/wiki/Implement-Queue-Service-Using-RocksDB

When doing a range scan in rocksdb, if there were keys recently deleted from that range, then you end up iterating over tombstones for all of those keys. In the case of a queue-like service, you may have many (potentially 100k to 1M+) tombstones.

Assuming it takes 2 hours for a tombstone to be fully compacted down to the bottom level (which is totally reasonable in the case of significant state > 20GB), if you have 50 tasks per second per partition then you can end up with 50 * 3600 * 2 == 360,000 tombstones in the range. This is because for every `TaskRun` we end up with a new `Schedule

In our Kafka Streams Processor `init()` method, we do a full prefix scan over `ScheduledTaskModel` store. In case of high per-partition load, this can take a very long time, causing the producer to get fenced which is a big orzdash.

This PR solves the problem with two changes:
* Storing `ScheduledTask`s in chronological order and processing them in chronological order.
* Remembering the timestamp of the last-processed ScheduledTask
* In the `init()` method, starting the iteration from the timestamp of the last processed `ScheduledTask` so as to skip over the tombstones in the range.