fix(server): properly map rocksdb log level (#1926)
We want at least `INFO_LEVEL` logs in RocksDB at the very least. This allows us to see the shape of the LSM tree, observe write stalls, and observe compactions; yet it is not very chatty. And it goes into `LOG` files on disk rather than stdout so it is cheaper.