fix(sdk-java): Close PollThread when receiving any exception during processing (#1928)
This PR improve PollThread by closing the thread instance after receiving RuntimeExceptions (or other throwables) when processing tasks. This is particularly because the RebalanceThread never detects that PollThread failed, causing worker instance to stuck.

We observed in our soak test environment that streamObserver.next() sometimes receives a ResourceLeakDetector$TraceRecord. I'm not sure if we need to take action to prevent this exception to happen, but at least we need to guard the PollThread from any possible RuntimeException.