feat(dashboard): Add task node run to the sidebard (#1950)
As users, you need to quickly understand the status of a task node run. Our current interface doesn't expose all the information we possess.

This change directly addresses that need:

New Data: This Pull Request (PR) adds the critical attempts and detailed status fields, allowing you to see the current state of the node.

Faster Loading: We are utilizing SWR (Stale-While-Revalidate) for data fetching. This means you will see the task information instantly from the cache, while the system updates the data seamlessly in the background.