fix(server): serialize checkpoints in output topic
Checkpointed Tasks were broken with the output topic enabled due to a silly bug. This commit stops shy of adding tests for the output topic to our e2e suite; however, it enables the output topic by default in our pipeline so that errors such as this would have been caught earlier.